Successfully build on Dec 8 2014 	VuiDeHoc Game v1.0.7 		+ Fixed time progress bar 		+ Changed Question Label

+++ MakeMathGame/Assets/Scripts/ResultNotifLabel.cs
@@ -16,7 +16,7 @@ public class ResultNotifLabel : MonoBehaviour {
 
 			}
 			correctNotifLabel.text = Program.score + "/10 câu";
-			timeNotifLabel.text = Program.playTime + " giây";
+			timeNotifLabel.text = (int)Program.playTime + " giây";
 
 
 		}else{
